requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
black==20.8b1
cfgv==3.2.0
click==7.1.2
decorator==5.0.7
distlib==0.3.1
Django==3.2
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.0
filelock==3.0.12
identify==2.2.4
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
mypy-extensions==0.4.3
nodeenv==1.6.0
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pre-commit==2.12.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.7.4.3
virtualenv==20.4.4
wcwidth==0.2.5
